requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
torch==2.0.1
torchvision==0.15.2
transformers==4.32.1

# MLOps tools
mlflow==2.5.0
dvc==3.12.0
feast==0.32.0
evidently==0.4.3
great-expectations==0.17.12

# Data processing
pyarrow==12.0.1
fastparquet==0.8.3
polars==0.18.15

# API and web framework
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.1.1
streamlit==1.25.0

# Database and storage
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
redis==4.6.0
minio==7.1.15

# Monitoring and logging
prometheus-client==0.17.1
grafana-api==1.0.3
loguru==0.7.0
structlog==23.1.0

# Kubernetes and deployment
kubernetes==27.2.0
istio-client==1.18.0
helm==3.12.2

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7
rich==13.5.2
typer==0.9.0

# Serialization
joblib==1.3.1
pickle5==0.0.12
cloudpickle==2.2.1

# HTTP clients
requests==2.31.0
httpx==0.24.1
aiohttp==3.8.5

# Image processing (for medical imaging models)
opencv-python==4.8.0.76
pillow==10.0.0
imageio==2.31.1

# Statistical analysis
scipy==1.11.1
statsmodels==0.14.0
seaborn==0.12.2
matplotlib==3.7.2
plotly==5.15.0

# Security and encryption
cryptography==41.0.3
pyjwt==2.8.0

# Performance optimization
numba==0.57.1
joblib==1.3.1

# Data validation
cerberus==1.3.5
marshmallow==3.20.1

# Feature engineering
feature-engine==1.6.2
category-encoders==2.6.1

# Model interpretation
shap==0.42.1
lime==0.2.0.1
eli5==0.13.0

# Time series (for longitudinal clinical data)
pmdarima==2.0.3
prophet==1.1.4

# Text processing (for clinical notes)
nltk==3.8.1
spacy==3.6.1
textblob==0.17.1

# Hyperparameter optimization
optuna==3.2.0
hyperopt==0.2.7
ray[tune]==2.6.3

# Model serving
bentoml==1.0.25
seldon-core==1.17.1

# Data drift detection
alibi-detect==0.11.4

# Workflow orchestration
prefect==2.11.3
airflow==2.7.0

# Environment management
conda==23.7.2
